Move `Placeholder` class to `execution` package (#3211)
* Move `Placeholder` class to `execution` package

It's not actually specific to PyTorch and is most closely related to
the other classes in `execution` (e.g. `Plans`, `Protocols`.)

* Remove stray `torch` import from `Placeholder`

* Point `syft-proto` dependency at the new release
