+++ pip-dep/requirements.txt
@@ -7,7 +7,7 @@ phe~=1.4.0
 Pillow~=6.2.2
 requests~=2.22.0
 scipy~=1.4.1
-syft-proto~=0.2.4.a1
+syft-proto~=0.2.5.a1
 tblib~=1.6.0
 torchvision~=0.5.0
 torch~=1.4.0
