+++ syft/serde/protobuf/proto.py
@@ -14,7 +14,7 @@ from syft.execution.plan import Plan
 from syft.execution.protocol import Protocol
 from syft.execution.state import State
 from syft.frameworks.torch.tensors.interpreters.additive_shared import AdditiveSharingTensor
-from syft.frameworks.torch.tensors.interpreters.placeholder import PlaceHolder
+from syft.execution.placeholder import PlaceHolder
 from syft.generic.pointers.pointer_tensor import PointerTensor
 from syft.messaging.message import TensorCommandMessage
 from syft.messaging.message import ObjectMessage
@@ -30,9 +30,7 @@ from syft_proto.execution.v1.state_pb2 import State as StatePB
 from syft_proto.frameworks.torch.tensors.interpreters.v1.additive_shared_pb2 import (
     AdditiveSharingTensor as AdditiveSharingTensorPB,
 )
-from syft_proto.frameworks.torch.tensors.interpreters.v1.placeholder_pb2 import (
-    Placeholder as PlaceholderPB,
-)
+from syft_proto.execution.v1.placeholder_pb2 import Placeholder as PlaceholderPB
 from syft_proto.generic.pointers.v1.pointer_tensor_pb2 import PointerTensor as PointerTensorPB
 from syft_proto.messaging.v1.message_pb2 import ObjectMessage as ObjectMessagePB
 from syft_proto.messaging.v1.message_pb2 import TensorCommandMessage as CommandMessagePB
