+++ test/serde/msgpack/test_msgpack_serde_full.py
@@ -70,7 +70,7 @@ samples[syft.federated.train_config.TrainConfig] = make_trainconfig
 samples[syft.workers.base.BaseWorker] = make_baseworker
 samples[syft.frameworks.torch.tensors.interpreters.autograd.AutogradTensor] = make_autogradtensor
 samples[syft.frameworks.torch.tensors.interpreters.private.PrivateTensor] = make_privatetensor
-samples[syft.frameworks.torch.tensors.interpreters.placeholder.PlaceHolder] = make_placeholder
+samples[syft.execution.placeholder.PlaceHolder] = make_placeholder
 samples[syft.frameworks.torch.fl.dataset.BaseDataset] = make_basedataset
 
 samples[syft.messaging.message.TensorCommandMessage] = make_command_message
